+++ lambda/sfr-publisher-reader/lib/readers/__init__.py
@@ -1,1 +1,2 @@
+from .iaReader import IAReader
 from .metReader import MetReader